requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
anthropic>=0.34.0
apscheduler==3.10.4
python-multipart==0.0.9
aiofiles==23.2.1
